requirements.txt
Django==5.2.7
mysqlclient==2.2.0
python-dotenv==1.0.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
unidecode==1.4.0
Faker==37.12.0
